consent-ws/src/test/java/org/broadinstitute/consent/http/mail/message/DisabledDatasetMessageTest.java
package org.broadinstitute.consent.http.mail.message;

import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import javax.mail.MessagingException;
import javax.mail.internet.MimeMessage;
import java.io.Writer;

import static org.junit.Assert.assertTrue;

public class DisabledDatasetMessageTest extends SessionHolder{

    @Mock
    Writer template;

    @Before
    public void setUp(){
        MockitoAnnotations.initMocks(this);
    }

    @Test
    public void testMessageSubject() throws MessagingException {
        MimeMessage message = new DisabledDatasetMessage().disabledDatasetMessage(getSession(), template, "DAR-123", "SomeType");
        assertTrue(message.getSubject().equals("Datasets not available for Data Access Request Application id: DAR-123."));
    }

}